fix #14 (ugly way)
This isn't the best looking way to fix it. In my opinion, having ADDCVARS take a string with wildcards as opposed to a string just to check for occurrences would definitely be better. However, the best method would be changing CatVar so that all categorization is done by the file initializing it. That would take a lot of rewriting, but if it were implemented correctly, could clean up the code considerably and make adding new hacks easier without reducing speed too much.