Fixed CanShoot reporting false on weapon reload
Should fix a big annoying bug which has existed for ages. where the aimbot will not shoot with silent aim when reloading because the nextprimaryattack netvar reports that the next attack is when the reload finishes - this can be remedied by only updating nextattack if we have fired the weapon. This works with 99% of weapons, except flamethrowers - for some reason this change makes canshoot always return true. You might also want to check for flamethrowers and use the traditional method for them. Also i'm not sure if this code even compiles since I don't have the bandwidth to clone your repo right now, but i'm sure if it doesn't it should be an easy fix.