Removed redundant screenshot of settings, others
Since the config file is now a requested item, we no longer need to ask the user for just literal screenshot of their menu in order to see what they have enabled.

Config is very useful as there may be other modules that can cause a crash when used in combination with another module. Removed "you can remove this line", making it a requirement.

Combined "drag and drop" information together, since it goes hand-in-hand. Made separate information in it's own separate block.

Please replace "paste symbolize-crash-log here instead" with "paste symbolize-crash-log here as well" if both symbolize-crash-log and logs are requested.